tic-tac-toe/src/main/java/Board.java
public class Board {
    private char[][] board;

    public Board() {
        board = new char[3][3];
        initializeBoard();
    }

    private void initializeBoard(){
        for(int i = 0; i < 3; i++){
            for(int j = 0; j < 3; j++){
                board[i][j] = '.';
            }
        }
    }

    public void moveBoard(int xPosition, int yPosition, Player player){
        board[xPosition][yPosition] = player.getMark();
    }

    public void printBoard(){
        for(int i = 0; i < 3; i++){
            for(int j = 0; j < 3; j++){
                System.out.print(board[i][j] + " ");
            }
            System.out.println();
        }
    }

    public char[][] getBoard() {
        return board;
    }

    public boolean isBoardFull() {
        for(int i = 0; i < 3; i++){
            for(int j = 0; j < 3; j++){
                if(board[i][j] == '.')
                    return false;
            }
        }
        return true;
    }
}
